requirements.txt
torch==2.9.1
torchvision==0.24.1
torchaudio==2.9.1
transformers==4.31.0
adapter-transformers==3.2.1
datasets==2.14.3
evaluate==0.4.0
accelerate==0.21.0
scikit-learn==1.3.0
pandas==2.0.3
numpy>=1.24.3
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
wandb==0.15.8
scipy==1.11.1
tqdm==4.65.0
pyyaml==6.0
tensorboard==2.13.0
jupyter==1.0.0
ipywidgets==8.0.6